Performance: Reduced Recycle Bin access time by 90% Recycle Bin support for Linux Performance: Reduced binary comparison sequential read time (by up to 75% for CD/DVD access) Improved synchronization sequence to avoid disk space shortage: overwrite large files by small ones first Fixed problems with file renaming on Samba share New free text grid search via shortcuts CTRL + F and F3 Show number of processed files at end of synchronization New optional grid column: file extension New comparison category icons Fixed handling sync-config of first folder pair Allow moving main dialog by dragging client area instead of title bar only Enhanced help file: Run RealTimeSync as Service Prefix log files with name of batch job Fixed GUI right-to-left mirroring for locales Hebrew and Arabic Portable version: save configuration in installation folder Many small GUI enhancements Updated translation files New Linux .deb package: ppa:freefilesync/ffs

+++ shared/customButton.cpp
@@ -1,3 +1,9 @@
+// **************************************************************************
+// * This file is part of the FreeFileSync project. It is distributed under *
+// * GNU General Public License: http://www.gnu.org/licenses/gpl.html       *
+// * Copyright (C) 2008-2010 ZenJu (zhnmju123 AT gmx.de)                    *
+// **************************************************************************
+//
 #include "customButton.h"
 #include <wx/dcmemory.h>
 #include <wx/image.h>
@@ -149,7 +155,7 @@ wxBitmap wxButtonWithImage::createBitmapFromText(const wxString& text)
     if ((accelPos = text.find(wxT("&"))) != wxString::npos)
     {
         textLabelFormatted.Replace(wxT("&"), wxT(""), false); //remove accelerator
-        indexAccel = accelPos;
+        indexAccel = static_cast<int>(accelPos);
     }
 
     dc.SetTextForeground(textColor);
